requirements.txt
beautifulsoup4==4.10.0
Flask==1.1.2
HeapDict==1.0.1
requests==2.26.0
gunicorn
lxml
pycryptodome